Implements user linking via code
Adds functionality to link Telegram users to school IDs using a code.
This allows users to access voting features after verification.

Handles cases where the user is not found, already linked, or the code is invalid,
providing informative error messages.

Adds an is_linked method to check if a user is linked or not
Adds SchoolUserNotFoundError and SchoolUserAlreadyLinkedError exceptions.